reload: Move initialize to new reload_setup module.
